+++ clease/corr_func.py
@@ -1,7 +1,7 @@
 """Module for calculating correlation functions."""
+from typing import Iterator, Tuple, Dict, Any
 import logging
 from ase.atoms import Atoms
-from ase.db import connect
 
 from clease_cxx import PyCEUpdater
 from .settings import ClusterExpansionSettings
@@ -11,61 +11,69 @@ from . import db_util
 logger = logging.getLogger(__name__)
 __all__ = ("CorrFunction", "ClusterNotTrackedError")
 
+# Type alias for a Correlation function
+CF_T = Dict[str, float]
+
 
 class ClusterNotTrackedError(Exception):
     """A cluster is not being tracked"""
 
 
 class CorrFunction:
-    """Calculate the correlation function.
+    """Class for calculating the correlation functions.
 
     Parameters:
 
-    settings: settings object
-
-    parallel: bool (optional)
-        specify whether or not to use the parallel processing for `get_cf`
-        method.
-
-    num_core: int or "all" (optional)
-        specify the number of cores to use for parallelization.
+        settings (ClusterExpansionSettings): The settings object which defines the
+            cluster expansion parameters.
     """
 
-    def __init__(self, settings):
-        if not isinstance(settings, ClusterExpansionSettings):
-            raise TypeError("setting must be CEBulk or CECrystal " "object")
+    def __init__(self, settings: ClusterExpansionSettings):
         self.settings = settings
 
-    def get_cf(self, atoms):
+    @property
+    def settings(self) -> ClusterExpansionSettings:
+        return self._settings
+
+    @settings.setter
+    def settings(self, value: Any) -> None:
+        if not isinstance(value, ClusterExpansionSettings):
+            raise TypeError(f"Setting must be a ClusterExpansionSettings object, got {value!r}")
+        self._settings = value
+
+    def connect(self, **kwargs):
+        return self.settings.connect(**kwargs)
+
+    def get_cf(self, atoms) -> CF_T:
         """
         Calculate correlation functions for all possible clusters and return
         them in a dictionary format.
 
         Parameters:
 
-        atoms: Atoms object
+            atoms (Atoms): The atoms object
         """
         if not isinstance(atoms, Atoms):
             raise TypeError("atoms must be an Atoms object")
         cf_names = self.settings.all_cf_names
         return self.get_cf_by_names(atoms, cf_names)
 
-    def get_cf_by_names(self, atoms, cf_names):
+    def get_cf_by_names(self, atoms, cf_names) -> CF_T:
         """
         Calculate correlation functions of the specified clusters and return
         them in a dictionary format.
 
         Parameters:
 
-        atoms: Atoms object
+            atoms: Atoms object
 
-        cf_names: list
-            names of correlation functions that will be calculated for
-            the structure provided in atoms
+            cf_names: list
+                names of correlation functions that will be calculated for
+                the structure provided in atoms
         """
 
         if isinstance(atoms, Atoms):
-            self.check_cell_size(atoms)
+            self.set_template(atoms)
         else:
             raise TypeError("atoms must be Atoms object")
 
@@ -77,50 +85,88 @@ class CorrFunction:
         cf = updater.calculate_cf_from_scratch(atoms, cf_names)
         return cf
 
+    def reconfigure_single_db_entry(self, row_id: int) -> None:
+        """Reconfigure a single DB entry. Assumes this is the initial structure,
+        and will not check that.
+
+        Parameters:
+
+            row_id: int
+                The ID of the row to be reconfigured.
+        """
+        with self.connect() as db:
+            atoms = wrap_and_sort_by_position(db.get(id=row_id).toatoms())
+            cf = self.get_cf(atoms)
+            db_util.update_table(db, row_id, self.cf_table_name, cf)
+
+    @property
+    def cf_table_name(self) -> str:
+        """Name of the table which holds the correlation functions."""
+        return f"{self.settings.basis_func_type.name}_cf"
+
+    def clear_cf_table(self) -> None:
+        """Delete the external table which holds the correlation functions."""
+        with self.connect() as db:
+            db.delete_external_table(self.cf_table_name)
+
     def reconfigure_db_entries(self, select_cond=None, verbose=False):
         """Reconfigure the correlation function values of the entries in DB.
 
         Parameters:
 
-        select_cond: list
-            -None (default): select every item in DB with
-                             "struct_type='initial'"
-            -else: select based on the condictions provided
-                  (struct_type='initial' is not automatically included)
+            select_cond: One of either:
 
-        verbose: bool
-            print the progress of reconfiguration if set to *True*
+                - None (default): select every item in DB with ``struct_type='initial'``
+                - Select based on the condictions provided (``struct_type='initial'`` is
+                  not automatically included)
+
+            verbose (bool):
+                print the progress of reconfiguration if set to *True*
         """
-        tab_name = "{}_cf".format(self.settings.basis_func_type.name)
-        db = connect(self.settings.db_name)
-        db.delete_external_table(tab_name)
-        select = []
-        if select_cond is not None:
-            for cond in select_cond:
-                select.append(cond)
-        else:
-            select = [("struct_type", "=", "initial")]
+        select = format_selection(select_cond)
+        db = self.connect()
 
         # get how many entries need to be reconfigured
-        row_ids = [row.id for row in db.select(select)]
-        num_reconf = len(row_ids)
+        num_reconf = db.count(select)
         msg = f"{num_reconf} entries will be reconfigured"
         logger.info(msg)
         if verbose:
             print(msg)
-        for count, row_id in enumerate(row_ids):
 
-            msg = f"updating {count+1} of {num_reconf} entries"
+        for row_id, count, total in self.iter_reconfigure_db_entries(select_cond=select_cond):
+            msg = f"Updated {count} of {total} entries. Current ID: {row_id}"
             if verbose:
                 print(msg, end="\r")
             logger.debug(msg)
-            atoms = wrap_and_sort_by_position(db.get(id=row_id).toatoms())
-            cf = self.get_cf(atoms)
-            db_util.update_table(db, row_id, tab_name, cf)
 
         if verbose:
             print("\nreconfiguration completed")
         logger.info("Reconfiguration complete")
+
+    def iter_reconfigure_db_entries(self, select_cond=None) -> Iterator[Tuple[int, int]]:
+        """Iterator which reconfigures the correlation function values in the DB,
+        which yields after each reconfiguration and reports on the progress.
+
+        For more information, see :py:meth:`~reconfigure_db_entries`.
+
+        Yields:
+            Tuple[int, int, int]: (row_id, count, total) A tuple containing the ID
+            of the row which was just reconfigured, current
+            count which has been reconfigured, as well as the total number of
+            reconfigurations which will be performed.
+            The percentage-wise progress is thus (count / total) * 100.
+        """
+        # Setup, ensure CF's are cleared first
+        self.clear_cf_table()
+        db = self.connect()
+        select = format_selection(select_cond)
+        # Fetch the total number of reconfigures
+        total = db.count(select)
+        for count, row in enumerate(db.select(select), start=1):
+            row_id = row.id
+            self.reconfigure_single_db_entry(row_id)
+            # Yield the current progress
+            yield row_id, count, total
 
     def reconfigure_inconsistent_cf_table_entries(self):
         """Find and correct inconsistent correlation functions in table."""
@@ -144,10 +190,10 @@ class CorrFunction:
         """Get IDs of the structures with inconsistent correlation functions.
 
         Note: consisent structures have the exactly the same list of cluster
-              names as stored in settings.cf_names.
+        names as stored in settings.cf_names.
         """
-        db = connect(self.settings.db_name)
-        tab_name = f"{self.settings.basis_func_type.name}_cf"
+        db = self.connect()
+        tab_name = self.cf_table_name
         cf_names = sorted(self.settings.all_cf_names)
         inconsistent_ids = []
         for row in db.select("struct_type=initial"):
@@ -168,16 +214,16 @@ class CorrFunction:
             logger.info("'%s' table has no inconsistent entries.", tab_name)
         return inconsistent_ids
 
-    def check_cell_size(self, atoms):
-        """Check the size of provided cell and create a template if necessary.
+    def set_template(self, atoms: Atoms) -> None:
+        """Check the size of provided cell and set as the currently active
+        template in the settings object.
 
         Parameters:
 
-        atoms: Atoms object
-            Unrelaxed structure
+            atoms (Atoms):
+                Unrelaxed structure
         """
         self.settings.set_active_template(atoms=atoms)
-        return atoms
 
     def _cf_name_exists(self, cf_name):
         """Return True if cluster name exists. Otherwise False.
@@ -197,3 +243,15 @@ class CorrFunction:
                 "ClusterExpansionSettings. Check that the cutoffs are "
                 "correct, and try to run reconfigure_settings"
             )
+
+
+def format_selection(select_cond=None, default_struct_type="initial"):
+    """DB selection formatter. Will default to selecting
+    all initial structures if None is specified."""
+    select = []
+    if select_cond is not None:
+        for cond in select_cond:
+            select.append(cond)
+    else:
+        select = [("struct_type", "=", default_struct_type)]
+    return select